adding error message for non existing instances

+++ resource/tasks.py
@@ -41,29 +41,46 @@ def createTask():
 def getTask(id):
   db = get_db()
   user_id = get_jwt_identity()
+  error = None
   task = db.execute('SELECT * FROM task WHERE user_id = ? AND id = ?', (user_id,id)).fetchone()
-  return jsonify(dict(task)), 200
+  if task is None:
+    error = "This task doesn't exist"
+  if error is None:
+    return jsonify(dict(task)), 200
+  return jsonify({'error': True, 'message': error}), 401
 
 @bp.route('/<int:id>', methods=['DELETE'])
 @jwt_required
 def deleteTask(id):
   db = get_db()
   user_id = get_jwt_identity()
-  db.execute('DELETE FROM task WHERE user_id = ? AND id = ?', (user_id,id))
-  db.commit()
-  return jsonify({"success": True}), 200
+  error = None
+  task = db.execute('SELECT * FROM task WHERE user_id = ? AND id = ?', (user_id,id)).fetchone()
+  if task is None:
+    error = "This task doesn't exist"
+  if error is None:
+    db.execute('DELETE FROM task WHERE user_id = ? AND id = ?', (user_id,id))
+    db.commit()
+    return jsonify({"success": True}), 200
+  return jsonify({'error': True, 'message': error}), 401
 
 @bp.route('/<int:id>', methods=['PUT'])
 @jwt_required
 def editTask(id):
   db = get_db()
   user_id = get_jwt_identity()
+  error = None
   if not request.json['title']:
       error = 'Title is required.'
   elif not request.json['body']:
       error = 'Body is required.'
-  db.execute('UPDATE task SET title = ?, body = ?WHERE user_id = ? AND id = ?', (request.json['title'],request.json['body'],user_id,id))
-  db.commit()
   task = db.execute('SELECT * FROM task WHERE user_id = ? AND id = ?', (user_id,id)).fetchone()
-  return jsonify(dict(task)), 200
+  if task is None:
+    error = "This task doesn't exist"
+  if error is None:
+    db.execute('UPDATE task SET title = ?, body = ?WHERE user_id = ? AND id = ?', (request.json['title'],request.json['body'],user_id,id))
+    db.commit()
+    task = db.execute('SELECT * FROM task WHERE user_id = ? AND id = ?', (user_id,id)).fetchone()
+    return jsonify(dict(task)), 200
+  return jsonify({'error': True, 'message': error}), 401
